Storage: Move lock package into search package
The lock package under pkg/storage/unified/search/lock/ has no external
importers and is only intended to be used by RemoteIndexStore. Moving it
into the parent search package is a precursor to exposing locking via
RemoteIndexStore.

All identifiers in the lock package were already unexported and there
are no name collisions with the search package, so this is a mechanical
move: rename files with a 'lock_' prefix and change 'package lock' to
'package search'.
