+++ pkg/storage/unified/search/lock_cdk_backend.go
@@ -1,4 +1,4 @@
-package lock
+package search
 
 import (
 	"bytes"
