+++ pkg/storage/unified/search/lock_cdk_backend_options.go
@@ -1,4 +1,4 @@
-package lock
+package search
 
 import (
 	"context"
