+++ pkg/storage/unified/search/lock_objstore.go
@@ -1,4 +1,4 @@
-package lock
+package search
 
 import (
 	"context"
